requirements.txt
APScheduler==3.9.1
Django==4.1.1
djangorestframework==3.13.1
requests==2.28.1
python-dotenv==0.21.0